sentiment_analysis_api/requirements.txt
Flask==3.1.3
langchain==0.3.27
langchain-community==0.3.28
langchain-groq==0.3.8
groq==0.30.0
python-dotenv==1.0.1
requests==2.32.5
pydantic==2.10.6
pydantic-core==2.27.2
